refactor: StringCalculator 코드 리팩토링
